Reorganize with repo module
